Split out variant code into several new sub-modules
